rcc-ssrl/src/training/requirements.txt
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0
webdataset>=0.2.100,<1.0
braceexpand>=0.1.7
pyarrow>=15.0.2
pandas>=2.3.3
scikit-learn>=1.4.2
pyyaml>=6.0
matplotlib>=3.9.0
seaborn>=0.13.0
mlflow>=3.1.0
